+++ src/unix/utils.c
@@ -3,7 +3,7 @@
 
 char *processx__tmp_string(SEXP str, int i) {
   const char *ptr = CHAR(STRING_ELT(str, i));
-  char *cstr = R_alloc(1, strlen(ptr) + 1);
+  char *cstr = R_alloc(1, (int) strlen(ptr) + 1);
   strcpy(cstr, ptr);
   return cstr;
 }
@@ -12,7 +12,7 @@ char **processx__tmp_character(SEXP chr) {
   size_t i, n = LENGTH(chr);
   char **cchr = (void*) R_alloc(n + 1, sizeof(char*));
   for (i = 0; i < n; i++) {
-    cchr[i] = processx__tmp_string(chr, i);
+    cchr[i] = processx__tmp_string(chr, (int) i);
   }
   cchr[n] = 0;
   return cchr;
